Removing main() method never used.
I also changed the way Swing objects are called. Just used the generated main that I'm removing on the "launcher class". See https://docs.oracle.com/javase/tutorial/uiswing/concurrency/dispatch.html.
